fix(route): 代码拆分优化
