+++ packages/taro-router/src/history.ts
@@ -20,3 +20,9 @@ export function setHistoryMode (mode?: 'hash' | 'browser', base = '/') {
 export function prependBasename (url = '') {
   return basename.replace(/\/$/, '') + '/' + url.replace(/^\//, '')
 }
+
+export const hasBasename = (path, prefix) =>
+  new RegExp('^' + prefix + '(\\/|\\?|#|$)', 'i').test(path)
+
+export const stripBasename = (path, prefix) =>
+  hasBasename(path, prefix) ? path.substr(prefix.length) : path
